Initial TS Support
Not sure why I try to upgrade so many big libraries to TS and ESM, it gets complicated quickly hehe.

My goal is that I want to have something like JSZip, but build using only vanilla library features, and to be fully ESM compatible, as well as have first-party type definitions. Also, using a Promise-based API is key as well.

https://stackoverflow.com/questions/14527820/does-html-5s-blobbuilder-still-work-in-google-chrome
https://gist.github.com/eligrey/1079572/aeac96e31f03bb8aeecedd5e5a6a89ac92c106cc
https://www.google.com/search?q=require%28%27setimmediate%27%29
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/concat
https://stackoverflow.com/questions/29676635/convert-uint8array-to-array-in-javascript