More Util Typings
I'm considering whether I should do this with a fork of the library instead, and work on modernizing that instead of the original, which seems to be a few development iteration types back (ESM, TypeScript, browser-first).

https://github.com/telerik/jszip-esm

It seems like a lot of work for modernizing older projects like these are that they are for Node to start with, and there are a lot of polyfills set up to make it work for modern code, rather than the other way around. To be fair, that's the same issue I'm having with my own projects from a few years ago too, so it's just going to happen, and it's still just something we have to deal with. I guess it's good that the same language has been around long enough for it to grow so much to where that is a concern. People see that as a downside to the JS ecosystem, but it also really shows how much things have grown over the years.

I wish not as many new TS projects were OOP, it's still the same language under the hood, so I find it interesting how the typings for things in that manner seems to have brought non-FP concepts over to JS's fairly functional neighborhood (at least for packages, it seems like most user-facing APIs use functions where possible, and maybe the underlying implementation may use a few classes).